division by functions
Shortened code and created a function
