+++ 2_hilbertize.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python3
-
+from hilbertize_functions import final #function to create final files
 import matplotlib.pyplot as plt
 import argparse
 import os
@@ -31,15 +31,6 @@ parser.add_argument(
 )
 args = parser.parse_args()  # 2 arguments at all
 
-
-def final(file, val, pref):
-    max_data = max(map(max, val))
-    min_data = min(map(min, val))
-    data_color = [[int(255*(x-min_data)/(max_data-min_data))
-                   for x in l] for l in val]
-    with open(file+pref, 'wb') as fpng:
-        w = png.Writer(len(val[0]), len(val), greyscale=True)
-        w.write(fpng, data_color)
 
 
 for one_file in args.files:
