+++ hilbertize_serial.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python3
-
+from hilbertize_functions import final  # function to create final files
 import argparse
 import os
 import re
